flask-chatbot/requirements.txt
flask>=2.0.0
pillow>=8.0.0
openai>=0.27.0
python-dotenv>=0.19.0